# docs/swagger.yaml
basePath: /
definitions:
...
  files.Download:
    properties:
      expires_at:
        type: string
      file:
        $ref: '#/definitions/files.Record'
      url:
        type: string
    type: object
  files.Page:
    properties:
      items:
        items:
          $ref: '#/definitions/files.Record'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  files.Record:
    properties:
      checksum_sha256:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      etag:
        type: string
      id:
        type: string
      original_name:
        type: string
      size_bytes:
        type: integer
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
...
  httptransport.CreateDepartmentRequest:
    properties:
      code:
        maxLength: 63
        type: string
      name:
        maxLength: 512
        type: string
      parent_id:
        type: string
      sort_order:
        type: integer
    required:
    - code
    - name
    type: object
...
  httptransport.DeleteDepartmentRequest:
    properties:
      id:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
  httptransport.DeleteFileRequest:
    properties:
      id:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
...
  httptransport.DepartmentIDRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  httptransport.DepartmentTreeRequest:
    properties:
      keyword:
        maxLength: 256
        type: string
    type: object
  httptransport.FileIDRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  httptransport.FilePageRequest:
    properties:
      content_types:
        items:
          type: string
        type: array
      created_at_from:
        type: string
      created_at_to:
        type: string
      created_by_ids:
        items:
          type: string
        type: array
      ids:
        items:
          type: string
        type: array
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      size_from:
        type: integer
      size_to:
        type: integer
    type: object
...
  httptransport.LoginRequest:
    properties:
      client_id:
        maxLength: 256
        type: string
      client_secret:
        maxLength: 4096
        type: string
    required:
    - client_id
    - client_secret
    type: object
  httptransport.LoginResponseBody:
    properties:
      access_token:
        type: string
      token_type:
        type: string
    type: object
...
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
...
  httptransport.SetDepartmentMembersRequest:
    properties:
      department_id:
        type: string
      members:
        items:
          $ref: '#/definitions/tenant.DepartmentMemberAssignment'
        maxItems: 1000
        type: array
    required:
    - department_id
    type: object
...
  httptransport.UpdateDepartmentRequest:
    properties:
      id:
        type: string
      name:
        maxLength: 512
        type: string
      parent_id:
        type: string
      sort_order:
        type: integer
      version:
        type: integer
    required:
    - id
    - name
    - version
    type: object
...
  tenant.Department:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      sort_order:
        type: integer
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
  tenant.DepartmentMemberAssignment:
    properties:
      is_primary:
        type: boolean
      membership_id:
        type: string
    required:
    - membership_id
    type: object
  tenant.DepartmentNode:
    properties:
      children:
        items:
          $ref: '#/definitions/tenant.DepartmentNode'
        type: array
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      sort_order:
        type: integer
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
...
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Client credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.LoginResponseBody'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httptransport.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httptransport.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/httptransport.Response'
      summary: Issue a JWT access token and record a security event
      tags:
      - authentication
...
  /api/v1/files/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: File and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.DeleteFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Logically delete a file using optimistic locking
      tags:
      - files
  /api/v1/files/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: File ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.FileIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/files.Download'
              type: object
      security:
      - Bearer: []
      summary: Create a short-lived file download URL
      tags:
      - files
  /api/v1/files/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: File ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.FileIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/files.Record'
              type: object
      security:
      - Bearer: []
      summary: Get file metadata
      tags:
      - files
  /api/v1/files/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: File filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.FilePageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/files.Page'
              type: object
      security:
      - Bearer: []
      summary: Page accessible file metadata
      tags:
      - files
  /api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File content
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/files.Record'
              type: object
      security:
      - Bearer: []
      summary: Upload a file
      tags:
      - files
...
  /api/v1/tenant-departments/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Department
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.Department'
              type: object
      security:
      - Bearer: []
      summary: Create a tenant department
      tags:
      - tenant-departments
  /api/v1/tenant-departments/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Department
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.DeleteDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Delete an empty tenant department
      tags:
      - tenant-departments
  /api/v1/tenant-departments/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Department
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.DepartmentIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.Department'
              type: object
      security:
      - Bearer: []
      summary: Get a tenant department
      tags:
      - tenant-departments
  /api/v1/tenant-departments/members/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: Assignments
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.SetDepartmentMembersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Replace department members and primary-department flags
      tags:
      - tenant-departments
  /api/v1/tenant-departments/tree:
    post:
      consumes:
      - application/json
      parameters:
      - description: Search
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.DepartmentTreeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/tenant.DepartmentNode'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Return the tenant department tree
      tags:
      - tenant-departments
  /api/v1/tenant-departments/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Department
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/tenant.Department'
              type: object
      security:
      - Bearer: []
      summary: Update or move a tenant department
      tags:
      - tenant-departments
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"